fix(ivy): Renderer2 should not use a special injection fn
