+++ packages/core/src/ivy_switch/runtime/ivy_switch_on.ts
@@ -6,9 +6,10 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {injectChangeDetectorRef, injectElementRef, injectTemplateRef, injectViewContainerRef} from '../../render3/view_engine_compatibility';
+import {injectChangeDetectorRef, injectElementRef, injectRenderer2, injectTemplateRef, injectViewContainerRef} from '../../render3/view_engine_compatibility';
 
 export const R3_ELEMENT_REF_FACTORY = injectElementRef;
 export const R3_TEMPLATE_REF_FACTORY = injectTemplateRef;
 export const R3_CHANGE_DETECTOR_REF_FACTORY = injectChangeDetectorRef;
 export const R3_VIEW_CONTAINER_REF_FACTORY = injectViewContainerRef;
+export const R3_RENDERER2_FACTORY = injectRenderer2;
