+++ packages/core/src/ivy_switch/runtime/legacy.ts
@@ -18,14 +18,17 @@ export const R3_CHANGE_DETECTOR_REF_FACTORY__POST_NGCC__: FactoryFunction =
     ivyOn.R3_CHANGE_DETECTOR_REF_FACTORY;
 export const R3_VIEW_CONTAINER_REF_FACTORY__POST_NGCC__: FactoryFunction =
     ivyOn.R3_VIEW_CONTAINER_REF_FACTORY;
+export const R3_RENDERER2_FACTORY__POST_NGCC__: FactoryFunction = ivyOn.R3_RENDERER2_FACTORY;
 
 
 export const R3_ELEMENT_REF_FACTORY__PRE_NGCC__ = noopFactory;
 export const R3_TEMPLATE_REF_FACTORY__PRE_NGCC__ = noopFactory;
 export const R3_CHANGE_DETECTOR_REF_FACTORY__PRE_NGCC__ = noopFactory;
 export const R3_VIEW_CONTAINER_REF_FACTORY__PRE_NGCC__ = noopFactory;
+export const R3_RENDERER2_FACTORY__PRE_NGCC__ = noopFactory;
 
 export let R3_ELEMENT_REF_FACTORY = R3_ELEMENT_REF_FACTORY__PRE_NGCC__;
 export let R3_TEMPLATE_REF_FACTORY = R3_TEMPLATE_REF_FACTORY__PRE_NGCC__;
 export let R3_CHANGE_DETECTOR_REF_FACTORY = R3_CHANGE_DETECTOR_REF_FACTORY__PRE_NGCC__;
 export let R3_VIEW_CONTAINER_REF_FACTORY = R3_VIEW_CONTAINER_REF_FACTORY__PRE_NGCC__;
+export let R3_RENDERER2_FACTORY = R3_RENDERER2_FACTORY__PRE_NGCC__;
