+++ packages/core/test/bundling/hello_world_r2/bundle.golden_symbols.json
@@ -1254,6 +1254,9 @@
     "name": "R3_ELEMENT_REF_FACTORY$1"
   },
   {
+    "name": "R3_RENDERER2_FACTORY$1"
+  },
+  {
     "name": "R3_TEMPLATE_REF_FACTORY$1"
   },
   {
